pinpoint functional
Added damage taken modifier to pinpoint status, and modifiers are implemented in take_damage(). The old placeholder focus status has been inactivated. Battle groups need to receive a signal when focus is applied to make sure any other fighers with focus have it removed (only one focus may be active per battle group). The get_front_or_focus needs to be changed to to respond to the new status system. I should consider getting rid of the focus member variable in battle_group and just let it be tracked through the existence of a status.